Redis testing
Redis functionality in progress.
